+++ tests/PHPUnit/Integration/ReportTest.php
@@ -9,6 +9,7 @@
 namespace Piwik\Tests\Integration;
 
 use Piwik\API\Proxy;
+use Piwik\Container\StaticContainer;
 use Piwik\Plugin\Report;
 use Piwik\Plugins\ExampleReport\Reports\GetExampleReport;
 use Piwik\Plugins\Actions\Columns\ExitPageUrl;
@@ -16,8 +17,8 @@ use Piwik\Piwik;
 use Piwik\Metrics;
 use Piwik\Plugins\ExampleTracker\Columns\ExampleDimension;
 use Piwik\Plugins\Referrers\Columns\Keyword;
+use Piwik\Translation\Translator;
 use Piwik\WidgetsList;
-use Piwik\Translate;
 use Piwik\Menu\MenuReporting;
 use Piwik\Plugin\Manager as PluginManager;
 use Piwik\Tests\Framework\Fixture;
@@ -156,16 +157,16 @@ class ReportTest extends IntegrationTestCase
 
     public function test_getWidgetTitle_shouldReturnTranslatedTitleIfSet()
     {
-        Translate::loadAllTranslations();
+        $this->loadTranslations();
         $this->assertEquals('Page Titles Following a Site Search', $this->advancedReport->getWidgetTitle());
-        Translate::reset();
+        $this->resetTranslations();
     }
 
     public function test_getCategory_shouldReturnTranslatedCategory()
     {
-        Translate::loadAllTranslations();
+        $this->loadTranslations();
         $this->assertEquals('Goals', $this->advancedReport->getCategory());
-        Translate::reset();
+        $this->resetTranslations();
     }
 
     public function test_configureWidget_shouldNotAddAWidgetIfNoWidgetTitleIsSet()
@@ -545,5 +546,20 @@ class ReportTest extends IntegrationTestCase
     private function unloadAllPlugins()
     {
         PluginManager::getInstance()->unloadPlugins();
+    }
+
+    private function loadTranslations()
+    {
+        /** @var Translator $translator */
+        $translator = StaticContainer::get('Piwik\Translation\Translator');
+        $translator->addDirectory(PIWIK_INCLUDE_PATH . '/lang');
+        PluginManager::getInstance()->loadPluginTranslations();
+    }
+
+    private function resetTranslations()
+    {
+        /** @var Translator $translator */
+        $translator = StaticContainer::get('Piwik\Translation\Translator');
+        $translator->reset();
     }
 }