+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
     * `Piwik\Registry`: use the container instead
     * `Piwik\TaskScheduler`: use `Piwik\Scheduler\Scheduler` instead
     * `Piwik\ScheduledTask`: use `Piwik\Scheduler\Task` instead
+    * `Piwik\Translate`: use `Piwik\Translation\Translator` instead
 * The following deprecated methods have been removed:
     * `Piwik\Common::getContinentsList()`: use `Piwik\Intl\Data\Provider\RegionDataProvider::getContinentList()` instead
     * `Piwik\Common::getCountriesList()`: use `Piwik\Intl\Data\Provider\RegionDataProvider::getCountryList()` instead